wikipedia-scraper/requirements.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
cffi==1.11.5
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
kafka-python==1.4.3
lxml==4.2.3
parsel==1.5.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==18.0.0
queuelib==1.5.0
schedule==0.5.0
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
Twisted==18.4.0
w3lib==1.19.0
zope.interface==4.5.0
